+++ History.md
@@ -5,6 +5,7 @@ Unpublished
   * use callbacks rather than events for asychronous open method
   * documentation
   * return this in synchronous write methods
+  * added close and closeSync methods
 
 0.1.0 / Dec 09 2014
 ===================